chore(files): move shared FileEntry and FileEntryGrid into a mixin
